+++ NEWS.md
@@ -1,3 +1,9 @@
+### packageRank 0.8.0.9013
+
+- refactor checkPackage().
+- change result data type and content of validatePackage().
+
+
 ### packageRank 0.8.0.9012
 
 - re-enable checkPackage() in packageHistory().